requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --extra=dev --output-file=requirements.txt pyproject.toml
#
agate==1.6.3
    # via dbt-core
attrs==22.1.0
    # via
    #   jsonschema
    #   pytest
babel==2.10.3
    # via agate
black==22.8.0
    # via jobby (pyproject.toml)
certifi==2022.9.14
    # via requests
cffi==1.15.1
    # via dbt-core
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   dbt-core
colorama==0.4.5
    # via dbt-core
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
dbt-core==1.3.0rc1
    # via jobby (pyproject.toml)
dbt-extractor==0.4.1
    # via dbt-core
fonttools==4.37.3
    # via matplotlib
future==0.18.2
    # via parsedatetime
hologram==0.0.15
    # via dbt-core
idna==3.4
    # via
    #   dbt-core
    #   requests
iniconfig==1.1.1
    # via pytest
isodate==0.6.1
    # via
    #   agate
    #   dbt-core
jinja2==3.1.2
    # via dbt-core
jsonschema==3.2.0
    # via hologram
kiwisolver==1.4.4
    # via matplotlib
leather==0.3.4
    # via agate
logbook==1.5.3
    # via dbt-core
loguru==0.6.0
    # via jobby (pyproject.toml)
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
mashumaro[msgpack]==3.0.4
    # via dbt-core
matplotlib==3.6.0
    # via jobby (pyproject.toml)
minimal-snowplow-tracker==0.0.2
    # via dbt-core
msgpack==1.0.4
    # via mashumaro
mypy-extensions==0.4.3
    # via black
networkx==2.8.6
    # via
    #   dbt-core
    #   jobby (pyproject.toml)
numpy==1.23.3
    # via
    #   contourpy
    #   matplotlib
packaging==21.3
    # via
    #   dbt-core
    #   matplotlib
    #   pytest
parsedatetime==2.4
    # via agate
pathspec==0.9.0
    # via
    #   black
    #   dbt-core
pillow==9.2.0
    # via matplotlib
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pydantic==1.10.2
    # via jobby (pyproject.toml)
pydot==1.4.2
    # via jobby (pyproject.toml)
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
    #   pydot
pyrsistent==0.18.1
    # via jsonschema
pytest==7.1.3
    # via jobby (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   hologram
    #   matplotlib
python-slugify==6.1.2
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2022.2.1
    # via babel
pyyaml==6.0
    # via dbt-core
requests==2.28.1
    # via
    #   dbt-core
    #   jobby (pyproject.toml)
    #   minimal-snowplow-tracker
six==1.16.0
    # via
    #   agate
    #   isodate
    #   jsonschema
    #   leather
    #   minimal-snowplow-tracker
    #   python-dateutil
sqlparse==0.4.3
    # via dbt-core
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   black
    #   pytest
typing-extensions==4.3.0
    # via
    #   black
    #   dbt-core
    #   mashumaro
    #   pydantic
urllib3==1.26.12
    # via requests
werkzeug==2.2.2
    # via dbt-core

# The following packages are considered to be unsafe in a requirements file:
# setuptools
